requirements.txt
asgiref==3.8.1
Django==5.1.1
django-contrib-postgres==0.0.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
environ==1.0
pillow==11.0.0
psycopg2==2.9.9
sqlparse==0.5.1
tzdata==2024.1